Some Linux and Debug-build centered changes from last month

+++ src/game/client/tf/c_tf_player.h
@@ -297,7 +297,7 @@ private:
 public:
 
 	Vector				m_vecPlayerColor;	
-	CUtlVector<float> 	m_iCosmetics;
+	CUtlVector<int> 	m_iCosmetics;
 	
 	void UpdatePlayerAttachedModels( void );
 	void UpdatePartyHat( void );
